initial rework [ci skip]

+++ src/commands/music/stop.ts
@@ -1,4 +1,4 @@
-import { ChatInputCommandInteraction } from 'discord.js';
+import { ChatInputCommandInteraction, GuildMember } from 'discord.js';
 import { localizedString, useLocalizedString } from '../../helpers/localization/localizedString.js';
 import { PlayerCommand } from '../../models/discord.js';
 
@@ -15,31 +15,49 @@ export const Stop: PlayerCommand = {
 
   run: async (interaction: ChatInputCommandInteraction) => {
     const { localize } = useLocalizedString(interaction.locale);
-    if (!interaction.guildId) {
-      const genericError = localize('global:genericError');
-      logger(DefaultLoggerMessage.GuildIsNotDefined).error();
-      return interaction.reply({
-        content: genericError,
+
+    try {
+      if (!interaction.guildId || !interaction.guild) {
+        logger(DefaultLoggerMessage.GuildIsNotDefined).error();
+        return await interaction.reply({
+          content: localize('global:genericError'),
+          ephemeral: true,
+        });
+      }
+      const player = useDefaultPlayer();
+      const queue = player.nodes.get(interaction.guildId);
+
+      if (!queue?.isPlaying()) {
+        return await interaction.reply({
+          content: localize('global:noMusicCurrentlyPlaying'),
+          ephemeral: true,
+        });
+      }
+
+      const memberChannel = (interaction.member as GuildMember | null)?.voice.channel;
+      if (!memberChannel || memberChannel.id !== queue.channel?.id) {
+        return await interaction.reply({
+          content: localize('global:mustBeInSameVoiceChannel'),
+          ephemeral: true,
+        });
+      }
+
+      queue.delete();
+
+      await interaction.reply({
+        content: localize('global:musicStopped'),
+      });
+    } catch (error) {
+      if (error instanceof Error) {
+        logger(error).error();
+      } else {
+        logger(String(error)).error();
+      }
+      await interaction.reply({
+        content: localize('global:genericError'),
         ephemeral: true,
       });
     }
-    const player = useDefaultPlayer();
-    const queue = player.nodes.get(interaction.guildId);
-
-    if (!queue?.isPlaying()) {
-      const noMusicCurrentlyPlaying = localize('global:noMusicCurrentlyPlaying');
-      return interaction.reply({
-        content: noMusicCurrentlyPlaying,
-        ephemeral: true,
-      });
-    }
-
-    queue.tracks.clear();
-    queue.delete();
-    const response = localize('global:musicStopped');
-    return interaction.reply({
-      content: response,
-    });
   },
 };
